pyproject.toml
[project]
name = "pixelscale"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.2",
    "pillow>=11.1.0",
    "scikit-image>=0.25.1",
]
